==> notebooks/01-ETL.ipynb <==
# Clothes Size Predictor 🧥

### ➤ Import Libraries and Stuff

import os
import sys
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import logging
logging.basicConfig(level=logging.INFO, format='[%(levelname)s] %(message)s')
logger = logging.getLogger(__name__)

# Get the current working directory
current_dir = os.getcwd()

# Navigate to the project root
project_root = os.path.abspath(os.path.join(current_dir, '..'))
logger.info(f"✅ Libraries uploaded")

⎙ Import the helpers from /src

try:
    sys.path.append(os.path.join(project_root))
    from src.pipeline.data import DataProcessor
    from src.pipeline.cleaning import DataCleaner
    logger.info(f"✅ Libraries uploaded")
except Exception as e:
    logger.info(f"❌ Error loading libraries: {e}")

⏏ Import the Dataset

# Load the cleaned dataset
file_path = os.path.abspath(os.path.join(project_root, 'data', 'raw', 'clothing_info.csv'))
processor = DataProcessor(file_path)
df = processor.run_pipeline()

## ➤ Data Cleaning

# --- Checking the nulls and deleting it
clothes_deploy = df.copy() # --- We save a copy here

clothes_deploy.shape

cleaner = DataCleaner(clothes_deploy)
clothes_deploy = cleaner.run_pipeline()

### --- Checking outliers

cleaner.handle_outliers()

# --- We remove it
cleaner.handle_outliers(action="remove")

cleaner.handle_outliers()

clothes_deploy.shape

clothes_deploy = cleaner.run_pipeline()

## --- Export the new CSV

clothes_deploy.to_csv(f'{project_root}/data/processed/clothes_processed.csv', index=False)
logger.info("✅ File saved as 'clothes_processed.csv'.")
