==> notebooks/03-Model_Training_and_Preprocessing.ipynb <==
# Clothes Size Predictor

## Quick Note (PLEASE READ)

This is the new notebook where I did the feature engineering and the training model, here's why:

I had the time to check where my model failed, and I noticed that my model is failing precisely because of the problems I identified (feature overlap and class imbalance). The only way to get out of that error is to attack those problems directly. 

### So let's begin

# 1. Preparación e Importación
import os
import sys
import pandas as pd
import joblib
from sklearn.model_selection import train_test_split, GridSearchCV
from sklearn.linear_model import LogisticRegression
from sklearn.ensemble import RandomForestClassifier
from sklearn.model_selection import GridSearchCV
from sklearn.metrics import classification_report

# Añadimos el path del proyecto para importar módulos personalizados
PROJECT_PATH = os.path.abspath(os.path.join(os.pardir))
sys.path.append(PROJECT_PATH)
from src.pipeline.feature_engineering import *

# Cargamos el dataset
DATA_PATH = '../data/processed/clothes_processed.csv' 
df = pd.read_csv(DATA_PATH)

# FILTRADO CRÍTICO: Eliminamos las filas con muy pocos datos para evitar sesgos
df = df[df['size'] != 'XXL'].copy() 
print(f"Filas restantes después de eliminar XXL (support=14): {len(df)}")


# 2. Ejecutar el Pipeline de Transformación
TARGET_COL = 'size'
processor = FeatureEngineer(df, target_col=TARGET_COL) 
df_processed = processor.run_all_preprocessing() 

# 3. Separación de Variables X e y
X = df_processed[['weight', 'height', 'age']] 
y = df_processed[f'{TARGET_COL}_encoded']      

# 4. División de Datos
X_train, X_test, y_train, y_test = train_test_split(
    X, y, 
    test_size=0.2, 
    random_state=42, 
    stratify=y  
)

# Define the Random Forest model
rf_model = RandomForestClassifier(random_state=42, n_jobs=-1)

# Hyperparameter grid for Random Forest
param_grid = {
    # NNumber of trees in the forest
    'n_estimators': [100, 200], 

    # Maximum depth of the trees
    'max_depth': [5, 10, None],  # None allows nodes to expand until all leaves are pure

    # Class weighting (to continue combating imbalance)
    'class_weight': [None, 'balanced'] 
}

# GridSearchCV
grid_search_rf = GridSearchCV(
    estimator=rf_model,
    param_grid=param_grid,
    scoring='f1_weighted',
    cv=5,
    n_jobs=-1
)

print("\n Initializing grid search for Random Forest...")
grid_search_rf.fit(X_train, y_train)

# Save the best model
best_rf_model = grid_search_rf.best_estimator_
joblib.dump(best_rf_model, '../models/best_rf_model_optimized.pkl')

print(f"Mejores Parámetros: {grid_search_rf.best_params_}")
print(f"Mejor Score (F1-Weighted): {grid_search_rf.best_score_:.4f}")

encoder = joblib.load('../models/label_encoder.pkl') 
y_pred = best_rf_model.predict(X_test)

print("\n--- REPORT ---")
print(classification_report(y_test, y_pred, target_names=encoder.classes_))
