==> notebooks/05-TestingTheModel.ipynb <==
## Clothes Size Prediction - TESTING ⌂

# Import necessary libraries
import os
import sys
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import logging
logging.basicConfig(level=logging.INFO, format='[%(levelname)s] %(message)s')
logger = logging.getLogger(__name__)

# Get the current working directory
current_dir = os.getcwd()

# Navigate to the project root
project_root = os.path.abspath(os.path.join(current_dir, '..'))

# Import from /src
sys.path.append(os.path.join(project_root))
logger.info(f"✅ Libraries Uploaded")

# --- Import from /src/pipelines
from src.pipeline.inference_pipeline import InferencePipeline
logger.info(f"✅ Libraries Uploaded")

# Importing the model
model_path = os.path.abspath(os.path.join(project_root, 'models', 'best_model.pkl'))
encoder_path = os.path.abspath(os.path.join(project_root, 'models', 'label_encoder.pkl'))
logger.info(f"✅ Model Imported")

# Initialize inference system
inference = InferencePipeline(model_path=model_path, encoder_path=encoder_path)

# Example input
new_person = {
    "weight": 72,   # kg
    "age": 28,      # years
    "height": 180   # cm
}

# Get prediction
predicted_size = inference.predict(new_person)

new_person2 = {
    "weight": 69,   # kg
    "age": 16,      # years
    "height": 165   # cm
}

# Get prediction
predicted_size = inference.predict(new_person2)

new_person3 = {
    "weight": 70,   # kg
    "age": 20,      # years
    "height": 176   # cm
}

# Get prediction
predicted_size = inference.predict(new_person3)

new_person4 = {
    "weight": 80,   # kg
    "age": 28,      # years
    "height": 180   # cm
}

# Get prediction
predicted_size = inference.predict(new_person4)
